Jackpot (DE/GER)
Runs a program where you enter six numbers and it randomly pulls numbers until your tip matches the pulled numbers. The number of times it has to be repeated are counted in the counter variable. The execution time is also timed but I'm not sure how the time package works in that regard.

This program is in German!